server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pymysql==1.1.1
pydantic==2.9.0
python-dotenv==1.0.1
cryptography==43.0.0
